requirements.txt
Django==2.2.3
django-redis==4.10.0
django-rest-framework==0.1.0
djangorestframework==3.10.1
Jinja2==2.10.1
MarkupSafe==1.1.1
pytz==2019.1
redis==3.3.0
sqlparse==0.3.0
uritemplate==3.0.0
